requirements.txt
selenium==4.10.0
python-dotenv==0.10.1
schedule==1.2.0
requests==2.28.2